Transferred Bar chart for activities into a new component frontend
