+++ FrontEnd/src/pages/AdminPage/UserProfilesTable/ActivitiesBarChart.module.css
@@ -1,0 +1,8 @@
+.chart-container {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  height: 250px;
+  width: 400px;
+}
