+++ FrontEnd/src/pages/AdminPage/UserProfilesTable/ProfilesStatistics.jsx
@@ -5,25 +5,7 @@ import { Descriptions, Segmented, Select, DatePicker } from 'antd';
 import Loader from '../../../components/Loader/Loader';
 import css from './ProfilesStatistics.module.css';
 import React from 'react';
-import {
-  Chart as ChartJS,
-  CategoryScale,
-  LinearScale,
-  BarElement,
-  Title,
-  Tooltip,
-  Legend,
-} from 'chart.js';
-import { Bar } from 'react-chartjs-2';
-
-ChartJS.register(
-  CategoryScale,
-  LinearScale,
-  BarElement,
-  Title,
-  Tooltip,
-  Legend
-);
+import ActivitiesBarChart from './ActivitiesBarChart';
 
 const { Option } = Select;
 
@@ -50,11 +32,7 @@ function ProfilesStatistics() {
   const url = `${baseUrl}/api/admin/profiles/statistics/${
     queryParams.length ? `?${queryParams.join('&')}` : ''
   }`;
-  const activities_url = `${baseUrl}/api/admin/profiles/statistics-activities/`;
-
   const { data: statistics, error, isLoading } = useSWR(url, fetcher);
-  const { data: activities, error: activitiesError, isLoading: activitiesLoading } = useSWR(activities_url, fetcher);
-
   const items = statistics
     ? [
         {
@@ -87,38 +65,6 @@ function ProfilesStatistics() {
   const handleRangeChange = (value, dateString) => {
     setPeriodRange({ start_date: dateString[0], end_date: dateString[1] });
   };
-  const chartData = activities
-  ?{
-      labels: ['Виробники', 'Імпортери', 'Роздрібніки', 'HORECA', 'Інші'],
-      datasets: [
-          {
-              label: 'Типи компаній',
-              data: [
-                  activities.manufacturers_count,
-                  activities.importers_count,
-                  activities.retail_networks_count,
-                  activities.horeca_count,
-                  activities.others_count
-              ],
-              backgroundColor: [
-                  '#87f3b0',
-              ]
-          },
-      ]
-  }: { labels: [], datasets: [] };
-  const options = {
-      responsive: true,
-      plugins: {
-        legend: {
-          position: 'top',
-        },
-        title: {
-          display: true,
-          text: 'Статистика по типам компаній',
-        },
-      },
-       maintainAspectRatio: false,
-    };
   return (
       <div className={css['statistics-container']}>
           <p className={css['statistics-title']}>Статистика компаній</p>
@@ -184,20 +130,7 @@ function ProfilesStatistics() {
                   items={items}
               />
           )}
-          {activitiesLoading && (
-              <div className={css['loader-container']}>
-                  <Loader/>
-              </div>
-          )}
-          {activitiesError && (
-              <div className={css['error']}>Не вдалося отримати статистику компаній</div>
-          )}
-          {!activitiesLoading && !activitiesError && (
-                <div className={css['chart-container']}>
-                    <Bar options={options} data={chartData}/>
-                </div>
-            )
-          }
+          <ActivitiesBarChart/>
       </div>
   );
 }
