+++ FrontEnd/src/pages/AdminPage/UserProfilesTable/ProfilesStatistics.module.css
@@ -6,13 +6,6 @@
   max-width: 400px;
   padding: 30px;
 }
-.chart-container {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  height: 250px;
-}
 
 .statistics-title {
   font-size: 16px;
